requirements.txt
# Core
google-generativeai>=0.3.0
numpy>=1.24.0
pandas>=2.0.0

# Dashboard
streamlit>=1.29.0

# Testing
pytest>=7.4.0

# Utilities
python-dotenv>=1.0.0
